ranim_core::animation

Trait Placeable

pub trait Placeable: Animation {
    // Provided method
    fn at(self, offset_sec: f64) -> At<Self> { ... }
}

Capability for animation definitions that have not been fixed in parent time coordinates.

This trait is used to constrain the anims to be inserted into AnimSequence. Only anims those are not placed can be inserted into it.

Provided Methods

fn at(self, offset_sec: f64) -> At<Self>

Place this definition at an offset in its parent’s local time coordinates.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".