ranim_core

Module traits

Fundamental traits.

Re-exports

pub use crate::anchor::Aabb;

pub use crate::anchor::AabbPoint;

pub use crate::anchor::Locate;

pub use transform::ApplyTransform;

pub use transform::Diag;

pub use transform::NotSimilarity;

pub use transform::Rigid;

pub use transform::RotateTransform;

pub use transform::ScaleHint;

pub use transform::ScaleTransform;

pub use transform::ScaleTransformExt;

pub use transform::ScaleTransformStrokeExt;

pub use transform::ShiftTransform;

pub use transform::ShiftTransformExt;

pub use transform::Similarity;

pub use transform::TransformGroup;

pub use transform::Translation;

pub use transform::UniformScaleTransform;

Modules

transform

Transform related things

Macros

impl_interpolatable_for_int 🔒

impl_interpolatable_tuple 🔒

Traits

AlignSlice

Align a slice of items

Alignable

A trait for aligning two items

ArrangeSlice

A trait for arranging operations.

Color

A trait for items that have both fill color and stroke color

Discard

A trait for discarding a value.

Empty

A trait for items that can be empty

FillColor

A trait for items that have fill color

Interpolatable

A trait for interpolating to values

Opacity

A trait for items with opacity

Partial

A trait for items that can be displayed partially

PointsFunc

A trait for items that can apply points function.

StrokeColor

A trait for items that have stroke color

StrokeWidth

A trait for items have stroke width

With

A trait for mutating a value in place.